refactor(agent): rename stream-* execution files to semantic names
