+++ packages/agent/src/core/chat-agent.ts
@@ -1,6 +1,6 @@
 import type { Sink } from "@openomni/protocol";
 import type { ChatAgentConfig, ChatAgentInput, AgentResult, AgentEvent } from "./types";
-import { streamAgent } from "./execution/stream-engine";
+import { streamAgent } from "./execution/runner";
 
 export interface ChatAgentInstance {
   run(input: ChatAgentInput, sink?: Sink): Promise<AgentResult>;
