+++ packages/agent/test/core/execution/stream-dispatch.test.ts
@@ -13,22 +13,22 @@ import {
   replacePrompt,
   rewriteWriteback,
 } from "../../helpers/policy-decision";
-import { buildPolicyEngine } from "../../../src/core/execution/stream-policy-engine";
-import { buildTurn } from "../../../src/core/execution/stream-turn";
+import { buildPolicyEngine } from "../../../src/core/execution/policy-engine-builder";
+import { buildTurn } from "../../../src/core/execution/turn-prepare";
 import {
   createStreamRunState,
   type StreamAgentBase,
   type StreamRunState,
   type TurnArtifacts,
-} from "../../../src/core/execution/stream-state";
+} from "../../../src/core/execution/run-state";
 import {
   dispatchBudgetCheck,
   dispatchModelRequest,
   dispatchModelResponse,
   dispatchPreRun,
-} from "../../../src/core/execution/stream-policy-dispatch";
-import { dispatchWritebackCommit } from "../../../src/core/execution/stream-writeback-policy";
-import { handleCompact, handleError, handleStop } from "../../../src/core/execution/stream-flow";
+} from "../../../src/core/execution/lifecycle-dispatch";
+import { dispatchWritebackCommit } from "../../../src/core/execution/writeback-policy";
+import { handleCompact, handleError, handleStop } from "../../../src/core/execution/turn-outcome";
 
 function makeInput(): ChatAgentInput {
   return { messages: [{ role: "user", content: "hello" }] };
